birthday.php php8 fix ('' not longer =0)

+++ plugins/fabrik_element/birthday/birthday.php
@@ -117,7 +117,7 @@ class PlgFabrik_ElementBirthday extends PlgFabrik_Element
 
 				// $$$ rob - all this below is nice but ... you still need to set a default
 				$detailValue = '';
-				$year = JString::ltrim($year, '0');
+				$year = (int)JString::ltrim($year, '0');
 
 				if (FabrikWorker::isDate($value))
 				{
